re-organize server html stuff
