+++ server/index.js
@@ -6,9 +6,7 @@ const config = require('config')
 // open libraries
 const express = require('express')
 require('express-async-errors');
-const fs = require('fs');
 const path = require('path');
-const serialize = require('serialize-javascript');
 const compression = require('compression');
 const cookieParser = require('cookie-parser');
 const mongoose = require('mongoose');
@@ -19,7 +17,7 @@ const MongoStore = require('connect-mongo');
 const errorHandler = require('./global/handlers/errorHandler.js')
 const { passport } = require('./global/handlers/passportHandler.js');
 
-// in dev the build path points to web/dist, on prod it points to web/build
+// on dev the build path points to web/dist, on prod it points to web/build
 const buildPath = config.get('buildPath');
 
 // init app
@@ -71,21 +69,6 @@ let router = express.Router();
 require('./global/api/router')(router, app)
 
 app.use('/', router);
-
-// serve the react app index.html
-app.get('*', (req, res) => {
-  const indexHtmlPath = path.resolve(`${buildPath}/index.html`);
-  fs.readFile(indexHtmlPath, 'utf8', (err, indexHtml) => {
-    if(err) {
-      console.error('Something went wrong:', err);
-      return res.status(500).send('Oops, better luck next time!');
-    }
-    // inject current user into the html by replacing the __CURRENT_USER__ placeholder in the html file. Info: https://create-react-app.dev/docs/title-and-meta-tags#injecting-data-from-the-server-into-the-page
-    // use `serialize` library to eliminate risk of XSS attacks when embedding JSON in html. Info: https://medium.com/node-security/the-most-common-xss-vulnerability-in-react-js-applications-2bdffbcc1fa0
-    const indexHtmlWithData = indexHtml.replace('__CURRENT_USER__', req.user ? serialize(req.user, { isJSON: true }) : '');
-    return res.send(indexHtmlWithData);
-  });
-});
 
 // unified error handler
 app.use(errorHandler)
